forwarding: extract TCP MSS / tunnel-MTU helpers into mss.rs
Pure code-motion split of forwarding/mod.rs (#5650). Move the TCP MSS
clamp and tunnel/GRE outer-MTU derivation helpers (effective_tcp_mss,
native_gre_inner_mtu, native_gre_tcp_mss, tunnel_outer_mtu,
tunnel_tcp_mss, select_tcp_mss) into forwarding/mss.rs.

Functions are moved verbatim; no logic, signature, or hot-path change.
pub(super) becomes pub(in crate::afxdp) so the mod.rs glob re-export
keeps all call sites resolving.

Validation: cargo build clean (0 errors).
